msm: mdss: Do not call set rate for handoff enabled clocks
When continuous splash screen is enabled, DSI branch clocks
are never turned off in the bootloader and are handed off
during kernel bootup. When the DSI controller is probed, it
needs to vote for keeping these clocks enabled. However,
calling set rate may result in a screen blink. Remove the
set rate call to fix this issue.

CRs-Fixed: 506037
Change-Id: I46074ff93456899de6b0e36a185d00d16f5f762e
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>

Conflicts:
	drivers/video/msm/mdss/msm_mdss_io_8974.c